requirements.txt
aiohttp==3.8.1
fastapi==0.70.0
python-decouple==3.5
requests==2.26.0
uvicorn==0.15.0